requirements.txt
Flask==2.2.3
python-docx==0.8.11
PyPDF2==3.0.1
gTTS==2.2.3
